Extricated distortion correction out of RM base class.
