+++ osvr/RenderKit/Float2.h
@@ -1,0 +1,48 @@
+/** @file
+    @brief Header
+
+    @date 2016
+
+    @author
+    Sensics, Inc.
+    <http://sensics.com>
+
+*/
+
+// Copyright 2016 Sensics, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// 	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef INCLUDED_Float2_h_GUID_CEEB5015_2C4A_4CA9_B701_23AA4409A946
+#define INCLUDED_Float2_h_GUID_CEEB5015_2C4A_4CA9_B701_23AA4409A946
+
+// Internal Includes
+// - none
+
+// Library/third-party includes
+// - none
+
+// Standard includes
+#include <array>
+
+namespace osvr {
+namespace renderkit {
+
+    /// 2D float data, like a texture coordinate for example.
+    using Float2 = std::array<float, 2>;
+
+} // namespace renderkit
+} // namespace osvr
+
+#endif // INCLUDED_Float2_h_GUID_CEEB5015_2C4A_4CA9_B701_23AA4409A946
+
