+++ osvr/RenderKit/RenderManagerBase.cpp
@@ -25,6 +25,9 @@ Russ Taylor <russ@sensics.com>
 // Internal Includes
 #include "RenderManager.h"
 #include <RenderManagerBackends.h>
+#include "DistortionCorrectTextureCoordinate.h"
+#include "DistortionParameters.h"
+#include "UnstructuredMeshInterpolator.h"
 
 #ifdef RM_USE_D3D11
 #include "RenderManagerD3D.h"
@@ -1816,570 +1819,6 @@ namespace renderkit {
         return std::sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
     }
 
-    RenderManager::UnstructuredMeshInterpolator::UnstructuredMeshInterpolator(
-        const MonoPointDistortionMeshDescription& points,
-        int numSamplesX, int numSamplesY
-        ) : m_points(points), m_numSamplesX(numSamplesX),
-        m_numSamplesY(numSamplesY) {
-
-      // Construct and fill in the grid of nearby points that is used by the
-      // interpolation function to accelerate the search for the three
-      // nearest non-collinear points.
-      std::vector< MonoPointDistortionMeshDescription> ySet;
-      MonoPointDistortionMeshDescription empty;
-      for (size_t y = 0; y < m_numSamplesY; y++) {
-        ySet.emplace_back(empty);
-      }
-      for (size_t x = 0; x < m_numSamplesX; x++) {
-        m_grid.emplace_back(ySet);
-      }
-      
-      // Go through each point in the unstructured grid and insert its index
-      // into all grid elements that are within 1/4th (rounded up) of the
-      // total span of the grid from its normalized location.
-      int xHalfSpan = static_cast<int>(0.9 + (1.0/4.0)*0.5 * m_numSamplesX);
-      int yHalfSpan = static_cast<int>(0.9 + (1.0/4.0)*0.5 * m_numSamplesY);
-      for (size_t i = 0; i < points.size(); i++) {
-        int xIndex, yIndex;
-        if (getIndex(points[i][0][0], points[i][0][1], xIndex, yIndex)) {
-
-          // Get the range of locations to insert
-          int xMin = xIndex - xHalfSpan;
-          if (xMin < 0) { xMin = 0; }
-          int xMax = xIndex + xHalfSpan;
-          if (xMax >= m_numSamplesX) { xMax = m_numSamplesX - 1; }
-          int yMin = yIndex - yHalfSpan;
-          if (yMin < 0) { yMin = 0; }
-          int yMax = yIndex + yHalfSpan;
-          if (yMax >= m_numSamplesY) { yMax = m_numSamplesY - 1; }
-
-          // Insert this point into each of these locations.
-          for (int x = xMin; x <= xMax; x++) {
-            for (int y = yMin; y <= yMax; y++) {
-              m_grid[x][y].push_back(points[i]);
-            }
-          }
-        }
-      }
-    }
-
-    Float2
-    RenderManager::UnstructuredMeshInterpolator::interpolateNearestPoints(
-            float xN, float yN) {
-        Float2 ret = {};
-
-        // Look in the spatial-acceleration grid to see if we can
-        // find three points without having to search the entire set of
-        // points.
-        int xIndex, yIndex;
-        if (!getIndex(xN, yN, xIndex, yIndex)) {
-          return ret;
-        }
-        MonoPointDistortionMeshDescription points;
-        points = getNearestPoints(xN, yN, m_grid[xIndex][yIndex]);
-
-        // If we didn't get enough points from the acceleration
-        // structure, look in the whole points array
-        if (points.size() < 3) {
-          points = getNearestPoints(xN, yN, m_points);
-        }
-
-        // If we didn't get three points, just return the output of
-        // the first point we found.
-
-        if (points.size() < 3) {
-          float xNew = static_cast<float>(points[0][1][0]);
-          float yNew = static_cast<float>(points[0][1][1]);
-          ret[0] = xNew;
-          ret[1] = yNew;
-          return ret;
-        }
-
-        // Found three points -- interpolate them.
-        float xNew = static_cast<float>(interpolate(
-          points[0][0][0], points[0][0][1], points[0][1][0],
-          points[1][0][0], points[1][0][1], points[1][1][0],
-          points[2][0][0], points[2][0][1], points[2][1][0], xN,
-          yN));
-        float yNew = static_cast<float>(interpolate(
-          points[0][0][0], points[0][0][1], points[0][1][1],
-          points[1][0][0], points[1][0][1], points[1][1][1],
-          points[2][0][0], points[2][0][1], points[2][1][1], xN,
-          yN));
-        ret[0] = xNew;
-        ret[1] = yNew;
-        return ret;
-    }
-
-    MonoPointDistortionMeshDescription
-      RenderManager::UnstructuredMeshInterpolator::getNearestPoints(
-          float xN, float yN,
-          const MonoPointDistortionMeshDescription &points) {
-
-      MonoPointDistortionMeshDescription ret;
-
-      // Find the three non-collinear points in the mesh that are nearest
-      // to the normalized point we are trying to look up.  We start by
-      // sorting the points based on distance from our location, selecting
-      // the first two, and then looking through the rest until we find
-      // one that is not collinear with the first two (normalized dot
-      // product magnitude far enough from 1).  If we don't find such
-      // points, we just go with the values from the closest point.
-      typedef std::multimap<double, size_t> PointDistanceIndexMap;
-      PointDistanceIndexMap map;
-      //if (points.size() < 460 && points.size() > 0) std::cout << "XXX Sorting mesh of size " << points.size() << std::endl;
-      for (size_t i = 0; i < points.size(); i++) {
-        // Insertion into the multimap sorts them by distance.
-        map.insert(std::make_pair(
-          pointDistance(xN, yN, points[i][0][0], points[i][0][1]), i));
-      }
-
-      PointDistanceIndexMap::const_iterator it = map.begin();
-      size_t first = it->second;
-      it++;
-      size_t second = it->second;
-      it++;
-      size_t third = first;
-      while (it != map.end()) {
-        if (!nearly_collinear(points[first][0], points[second][0],
-          points[it->second][0])) {
-          third = it->second;
-          break;
-        }
-        it++;
-      }
-
-      // Push back all of the points we found, which may not include
-      // a third point if the first is the same as the third.
-      if (map.size() >= 1) { ret.push_back(points[first]); }
-      if (map.size() >= 2) { ret.push_back(points[second]); }
-      if (third != first) {
-        ret.push_back(points[third]);
-      }
-      return ret;
-    }
-
-    Float2 RenderManager::DistortionCorrectTextureCoordinate(
-        size_t eye //< Which eye?
-        , Float2 const& inCoords //< Coordinates to modify
-        , DistortionParameters distort //< Distortion parameters
-        , size_t color //< 0 = red, 1 = green, 2 = blue
-        ) {
-        Float2 ret = inCoords;
-
-        // Check for invalid parameters
-        if (color > 2) {
-            return ret;
-        }
-        using Eigen::Vector2f;
-        // Convert from coordinates in the overfilled texture to coordinates
-        // that will cover the range (0,0) to (1,1) on the screen.  This is
-        // done by scaling around (0.5,0.5) to push the edges of the screen
-        // out to the (0,0) and (1,1) boundaries.
-        Eigen::Map<Vector2f> retMap(ret.data());
-        Eigen::Map<const Vector2f> inMap(inCoords.data());
-
-        Vector2f xyN = (inMap - Vector2f::Constant(0.5f)) *
-                           m_params.m_renderOverfillFactor +
-                       Vector2f::Constant(0.5f);
-        /// these are for mono point that hasn't been ported to eigen yet.
-        float x = inCoords[0];
-        float y = inCoords[1];
-        float xN = xyN.x();
-        float yN = xyN.y();
-
-        switch (distort.m_type) {
-        case osvr::renderkit::RenderManager::DistortionParameters::
-            rgb_symmetric_polynomials: {
-            if (distort.m_distortionPolynomialRed.size() < 2) {
-                return ret;
-            }
-            if (distort.m_distortionPolynomialGreen.size() < 2) {
-                return ret;
-            }
-            if (distort.m_distortionPolynomialBlue.size() < 2) {
-                return ret;
-            }
-            if (distort.m_distortionCOP.size() != 2) {
-                return ret;
-            }
-            if (distort.m_distortionD.size() != 2) {
-                return ret;
-            }
-            if (distort.m_distortionD[0] <= 0) {
-                return ret;
-            }
-            if (distort.m_distortionD[1] <= 0) {
-                return ret;
-            }
-
-            // Convert from normalized range to (D[0], D[1]) range.  Here,
-            // both coordinate systems share a common (0,0) boundary so we
-            // can just scale around the origin.
-            Vector2f xyD =
-                xyN.cwiseProduct(Vector2f::Map(distort.m_distortionD.data()));
-
-            // Compute the distance from the COP in D space
-            // (direction and squared magnitude)
-            Vector2f xyDDiff =
-                xyD - Vector2f::Map(distort.m_distortionCOP.data());
-            float rMag2 = xyDDiff.squaredNorm();
-            if (rMag2 == 0) { // We're at the center -- no distortion
-                ret = inCoords;
-                return ret;
-            }
-            float rMag = sqrt(rMag2);
-            Vector2f xyDNorm = xyDDiff / rMag;
-
-            // Compute the new location in D space based on the distortion
-            // parameters
-            std::vector<float> const& params =
-                color == 0 ? distort.m_distortionPolynomialRed
-                           : color == 1 ? distort.m_distortionPolynomialGreen
-                                        : distort.m_distortionPolynomialBlue;
-
-            // The distance scaling factor needs to be applied as many times
-            // as the degree of the polynomial we are using.  For the constant
-            // term, the factor is 1.
-            float rFactor = 1;
-            float rNew = params[0];
-
-            auto n = params.size();
-            using index_type = decltype(n);
-            for (index_type i = 1; i < n; i++) {
-                rFactor *= rMag;
-                rNew += params[i] * rFactor;
-            }
-            Vector2f xyDNew =
-                Vector2f::Map(distort.m_distortionCOP.data()) + rNew * xyDNorm;
-
-            // Convert from D space back to unit space
-            Vector2f xyNNew = xyDNew.cwiseQuotient(
-                Vector2f::Map(distort.m_distortionD.data()));
-
-            Vector2f::Map(ret.data()) = xyNNew;
-        } break;
-        case osvr::renderkit::RenderManager::DistortionParameters::
-            mono_point_samples: {
-            if (eye >= distort.m_monoPointSamples.size()) {
-                return ret;
-            }
-
-            MonoPointDistortionMeshDescription& points =
-                distort.m_monoPointSamples[eye];
-            if (points.size() < 3) {
-                return ret;
-            }
-
-            // Find the three non-collinear points in the mesh that are nearest
-            // to the normalized point we are trying to look up.  We start by
-            // sorting the points based on distance from our location, selecting
-            // the first two, and then looking through the rest until we find
-            // one that is not collinear with the first two (normalized dot
-            // product magnitude far enough from 1).  If we don't find such
-            // points, we just go with the values from the closest point.
-            ret = m_interpolators[0]->interpolateNearestPoints(xN, yN);
-        } break;
-        case osvr::renderkit::RenderManager::DistortionParameters::
-            rgb_point_samples: {
-            if (color >= 3) {
-                return ret;
-            }
-            if (eye >= distort.m_rgbPointSamples[color].size()) {
-                return ret;
-            }
-
-            MonoPointDistortionMeshDescription& points =
-                distort.m_rgbPointSamples[color][eye];
-            if (points.size() < 3) {
-                return ret;
-            }
-
-            // Find the three non-collinear points in the mesh that are nearest
-            // to the normalized point we are trying to look up.  We start by
-            // sorting the points based on distance from our location, selecting
-            // the first two, and then looking through the rest until we find
-            // one that is not collinear with the first two (normalized dot
-            // product magnitude far enough from 1).  If we don't find such
-            // points, we just go with the values from the closest point.
-            ret = m_interpolators[color]->interpolateNearestPoints(xN, yN);
-        } break;
-        default:
-            break;
-        }
-
-        // Convert from unit (normalized) space back into overfill space.
-        ret[0] = (ret[0] - 0.5f) / m_params.m_renderOverfillFactor + 0.5f;
-        ret[1] = (ret[1] - 0.5f) / m_params.m_renderOverfillFactor + 0.5f;
-
-        return ret;
-    }
-
-    RenderManager::DistortionMesh
-    RenderManager::ComputeDistortionMesh(
-        size_t eye //< Which eye?
-        , DistortionMeshType type //< Type of mesh to produce
-        , DistortionParameters distort //< Distortion parameters
-        ) {
-        RenderManager::DistortionMesh ret;
-
-        // Clear any created interpolators, freeing up their memory
-        // first.  These may have been left behind by a failed
-        // mesh-creation from before.
-        for (size_t clr = 0; clr < m_interpolators.size(); clr++) {
-            delete m_interpolators[clr];
-        }
-        m_interpolators.clear();
-
-        // Check the validity of the parameters, based on the ones we're
-        // using.
-        if (distort.m_type ==
-            RenderManager::DistortionParameters::rgb_symmetric_polynomials) {
-            if (distort.m_distortionPolynomialRed.size() < 2) {
-                std::cerr << "RenderManager::ComputeDistortionMesh: Need 2+ "
-                             "red polynomial coefficients, found "
-                          << distort.m_distortionPolynomialRed.size()
-                          << std::endl;
-                return ret;
-            }
-            if (distort.m_distortionPolynomialGreen.size() < 2) {
-                std::cerr << "RenderManager::ComputeDistortionMesh: Need 2+ "
-                             "green polynomial coefficients, found "
-                          << distort.m_distortionPolynomialGreen.size()
-                          << std::endl;
-                return ret;
-            }
-            if (distort.m_distortionPolynomialBlue.size() < 2) {
-                std::cerr << "RenderManager::ComputeDistortionMesh: Need 2+ "
-                             "blue polynomial coefficients, found "
-                          << distort.m_distortionPolynomialBlue.size()
-                          << std::endl;
-                return ret;
-            }
-            if (distort.m_distortionD.size() != 2) {
-                std::cerr << "RenderManager::ComputeDistortionMesh: Need 2 "
-                             "distortion coefficients, found "
-                          << distort.m_distortionD.size() << std::endl;
-                return ret;
-            }
-        } else if (distort.m_type ==
-                   RenderManager::DistortionParameters::mono_point_samples) {
-            if (distort.m_monoPointSamples.size() != 2) {
-                std::cerr << "RenderManager::ComputeDistortionMesh: Need 2 "
-                             "meshes, found "
-                          << distort.m_monoPointSamples.size() << std::endl;
-                return ret;
-            }
-            // Add a new interpolator to be used when we're finding
-            // mesh coordinates.
-            if (distort.m_monoPointSamples[eye].size() < 3) {
-                std::cerr << "RenderManager::ComputeDistortionMesh: Need "
-                              "3+ points, found "
-                          << distort.m_monoPointSamples[eye].size()
-                          << std::endl;
-                return ret;
-            }
-            for (size_t i = 0; i < distort.m_monoPointSamples[eye].size();
-                  i++) {
-                if (distort.m_monoPointSamples[eye][i].size() != 2) {
-                    std::cerr
-                        << "RenderManager::ComputeDistortionMesh: Need 2 "
-                        << "points in the mesh, found "
-                        << distort.m_monoPointSamples[eye][i].size()
-                        << std::endl;
-                    return ret;
-                }
-                if (distort.m_monoPointSamples[eye][i][0].size() != 2) {
-                    std::cerr
-                        << "RenderManager::ComputeDistortionMesh: Need 2 "
-                        << "values in input point, found "
-                        << distort.m_monoPointSamples[eye][i][0].size()
-                        << std::endl;
-                    return ret;
-                }
-                if (distort.m_monoPointSamples[eye][i][1].size() != 2) {
-                    std::cerr
-                        << "RenderManager::ComputeDistortionMesh: Need 2 "
-                        << "values in output point, found "
-                        << distort.m_monoPointSamples[eye][i][1].size()
-                        << std::endl;
-                    return ret;
-                }
-            }
-            // Add a new interpolator to be used when we're finding
-            // mesh coordinates.
-            m_interpolators.emplace_back(new
-              UnstructuredMeshInterpolator(distort.m_monoPointSamples[eye]));
-        }
-        else if (distort.m_type ==
-                   RenderManager::DistortionParameters::rgb_point_samples) {
-            if (distort.m_rgbPointSamples.size() != 3) {
-                std::cerr << "RenderManager::ComputeDistortionMesh: Need 3 "
-                             "color meshes, found "
-                          << distort.m_rgbPointSamples.size() << std::endl;
-                return ret;
-            }
-            for (size_t clr = 0; clr < 3; clr++) {
-                if (distort.m_rgbPointSamples[clr].size() != 2) {
-                    std::cerr << "RenderManager::ComputeDistortionMesh: Need 2 "
-                                 "eye meshes, found "
-                              << distort.m_rgbPointSamples[clr].size()
-                              << std::endl;
-                    return ret;
-                }
-                // Add a new interpolator to be used when we're finding
-                // mesh coordinates.
-                if (distort.m_rgbPointSamples[clr][eye].size() < 3) {
-                    std::cerr
-                        << "RenderManager::ComputeDistortionMesh: Need "
-                            "3+ points, found "
-                        << distort.m_rgbPointSamples[clr][eye].size()
-                        << std::endl;
-                    return ret;
-                }
-                for (size_t i = 0;
-                      i < distort.m_rgbPointSamples[clr][eye].size(); i++) {
-                    if (distort.m_rgbPointSamples[clr][eye][i].size() !=
-                        2) {
-                        std::cerr
-                            << "RenderManager::ComputeDistortionMesh: Need "
-                                "2 "
-                            << "points in the mesh, found "
-                            << distort.m_rgbPointSamples[clr][eye][i].size()
-                            << std::endl;
-                        return ret;
-                    }
-                    if (distort.m_rgbPointSamples[clr][eye][i][0].size() !=
-                        2) {
-                        std::cerr
-                            << "RenderManager::ComputeDistortionMesh: Need "
-                                "2 "
-                            << "values in input point, found "
-                            << distort.m_rgbPointSamples[clr][eye][i][0]
-                                    .size()
-                            << std::endl;
-                        return ret;
-                    }
-                    if (distort.m_rgbPointSamples[clr][eye][i][1].size() !=
-                        2) {
-                        std::cerr
-                            << "RenderManager::ComputeDistortionMesh: Need "
-                                "2 "
-                            << "values in output point, found "
-                            << distort.m_rgbPointSamples[clr][eye][i][1]
-                                    .size()
-                            << std::endl;
-                        return ret;
-                    }
-                }
-
-                // Add a new interpolator to be used when we're finding
-                // mesh coordinates, one per eye.
-                m_interpolators.emplace_back(new
-                  UnstructuredMeshInterpolator(distort.m_rgbPointSamples[clr][eye]));
-            }
-        } else {
-            std::cerr << "RenderManager::ComputeDistortionMesh: Unrecognized "
-                      << "distortion parameter type" << std::endl;
-            return ret;
-        }
-
-        // See what kind of mesh we're supposed to produce.  Make the
-        // appropriate one.
-        switch (type) {
-        case SQUARE: {
-            // Figure out how many quads we should use in each dimension.  The
-            // minimum is 1.  We have an even number in each.  There are two
-            // triangles per quad.
-            int quadsPerSide =
-                static_cast<int>(sqrt(distort.m_desiredTriangles / 2));
-            if (quadsPerSide < 1) {
-                quadsPerSide = 1;
-            }
-
-            // Figure out how large each quad will be.  Recall that we're
-            // covering a range of 2 (from -1 to 1) in each dimension, so the
-            // quads will all be square in texture space.
-            float quadSide = 2.0f / quadsPerSide;
-            float quadTexSide = 1.0f / quadsPerSide;
-
-            // Compute distorted texture coordinates and use those for each
-            // vertex, with appropriate spatial location and texture
-            // coordinates.
-
-            auto const numVertsPerSide = quadsPerSide + 1;
-            auto const numVertices = numVertsPerSide*numVertsPerSide;
-            ret.vertices.reserve(numVertices);
-
-            // Generate a grid of vertices with distorted texture coordinates
-            for (int x = 0; x < numVertsPerSide; x++) {
-                float xPos = -1 + x * quadSide;
-                float xTex = x * quadTexSide;
-
-                for (int y = 0; y < numVertsPerSide; y++) {
-                    float yPos = -1 + y * quadSide;
-                    float yTex = y * quadTexSide;
-
-                    Float2 pos = { xPos, yPos };
-                    Float2 tex = { xTex, yTex };
-
-                    ret.vertices.emplace_back(pos,
-                        DistortionCorrectTextureCoordinate(eye, tex, distort, 0),
-                        DistortionCorrectTextureCoordinate(eye, tex, distort, 1),
-                        DistortionCorrectTextureCoordinate(eye, tex, distort, 2));
-                }
-            }
-
-            // Generate a pair of triangles for each quad, wound
-            // counter-clockwise from the mesh grid
-
-            // total of quadsPerSide * quadsPerSide * 6 vertices added: reserve
-            // that space to avoid excess copying during mesh generation.
-            ret.indices.reserve(quadsPerSide * quadsPerSide * 6);
-            for (int x = 0; x < quadsPerSide; x++) {
-                for (int y = 0; y < quadsPerSide; y++) {
-                    // Grid generated above is in column-major order
-                    int indexLL = x*numVertsPerSide + y;
-                    int indexHL = indexLL + numVertsPerSide;
-                    int indexHH = indexLL + numVertsPerSide + 1;
-                    int indexLH = indexLL + 1;
-
-                    // Triangle 1
-                    ret.indices.emplace_back(indexLL);
-                    ret.indices.emplace_back(indexHL);
-                    ret.indices.emplace_back(indexHH);
-
-                    // Triangle 2
-                    ret.indices.emplace_back(indexLL);
-                    ret.indices.emplace_back(indexHH);
-                    ret.indices.emplace_back(indexLH);
-                }
-            }
-        } break;
-        case RADIAL: {
-            std::cerr
-                << "RenderManager::ComputeDistortionMesh: Radial mesh type "
-                << "not yet implemented" << std::endl;
-
-            // @todo Scale the aspect ratio of the rings around the center of
-            // projection so that they will be round in the visible display.
-        } break;
-        default:
-            std::cerr << "RenderManager::ComputeDistortionMesh: Unsupported "
-                         "mesh type: "
-                      << type << std::endl;
-        }
-
-        // Clear any created interpolators, freeing up their memory
-        // first.
-        for (size_t clr = 0; clr < m_interpolators.size(); clr++) {
-            delete m_interpolators[clr];
-        }
-        m_interpolators.clear();
-
-        return ret;
-    }
-
     static std::string osvrRenderManagerGetString(OSVR_ClientContext context,
                                                   const std::string& path) {
         size_t len;
@@ -2699,7 +2138,7 @@ namespace renderkit {
           // Turn on distortion correction and set the parameters for
           // this eye.  We assume that the parameters are in the space
           // that is the size of the display.
-          RenderManager::DistortionParameters dP;
+          DistortionParameters dP;
           dP.m_desiredTriangles = 800;
           std::vector<float> COP;
           COP.push_back(static_cast<float>(rDP.centerOfProjection.data[0]));
@@ -2731,7 +2170,7 @@ namespace renderkit {
           p.m_distortionCorrection = true;
         } else {
           // Push back an empty set of parameters for this eye
-          p.m_distortionParameters.push_back(RenderManager::DistortionParameters());
+          p.m_distortionParameters.push_back(DistortionParameters());
         }
       }
     }
@@ -2740,12 +2179,12 @@ namespace renderkit {
         // Construct the distortion parameters based on the local display
         // class.
         // @todo Remove once we get a general polynomial from Core.
-        RenderManager::DistortionParameters distortion;
+        DistortionParameters distortion;
         distortion.m_desiredTriangles = 200 * 64;
 
         if (p.m_displayConfiguration.getDistortionType() ==
             OSVRDisplayConfiguration::RGB_SYMMETRIC_POLYNOMIALS) {
-            distortion.m_type = osvr::renderkit::RenderManager::
+            distortion.m_type = osvr::renderkit::
                 DistortionParameters::Type::rgb_symmetric_polynomials;
             std::vector<float> Ds;
             Ds.push_back(
@@ -2771,7 +2210,7 @@ namespace renderkit {
             }
         } else if (p.m_displayConfiguration.getDistortionType() ==
                    OSVRDisplayConfiguration::MONO_POINT_SAMPLES) {
-            distortion.m_type = osvr::renderkit::RenderManager::
+            distortion.m_type = osvr::renderkit::
                 DistortionParameters::Type::mono_point_samples;
             distortion.m_monoPointSamples =
                 p.m_displayConfiguration.getDistortionMonoPointMeshes();
@@ -2785,7 +2224,7 @@ namespace renderkit {
                 p.m_displayConfiguration.getEyes().size(), distortion);
         } else if (p.m_displayConfiguration.getDistortionType() ==
                    OSVRDisplayConfiguration::RGB_POINT_SAMPLES) {
-            distortion.m_type = osvr::renderkit::RenderManager::
+            distortion.m_type =
                 DistortionParameters::Type::rgb_point_samples;
             distortion.m_rgbPointSamples =
                 p.m_displayConfiguration.getDistortionRGBPointMeshes();
